iOS: Enable ignoring of extra scroll events
The change enables developers to ignore scroll events that occur between the time they set the scroll position in JavaScript and the time that the native code applies the scroll position to the view. This is important for apps that care about carefully tracking and managing the ScrollView's scroll position. For example, when a virtualized ScrollView sets the scroll position to X, it wants to begin preparing to render the content as though it's scrolled to X. Currently, if it receives a scroll event for Y, it doesn't know whether the ScrollView reached Y before or after X. Consequently, it doesn't know whether it should start rendering content for Y because Y is the latest scroll position or if it should ignore Y and continue rendering content for X.

The way this is implemented is each method for changing the scroll position (e.g. scrollTo, scrollToEnd) returns an id. When the native code sets the scroll position on the view, it fires an `onScrollPositionSet` which contains the id of the related scroll request. Apps can solve the problem described above by ignoring all scroll events that occur between the time that the app requests the scroll position change and the time the corresponding `onScrollPositionSet` event fires.

**Test Plan**

Built a test app that uses the above technique to ignore scroll events that occur between the scroll request and the firing of the corresponding `onScrollPositionSet` event. Logged whether each `onScroll` event should be ignored and verified that the log matched my expectations. Tested this with `scrollTo`, `scrollToEnd`, and `scrollBy`.

**Dependency Notes**

I built this change on top of my other PR (#15609). If I didn't do this, I'd have to resolve merge conflicts after #15609 was merged.

Adam Comella
Microsoft Corp.